+++ odoo/local-src/connector_esb/models/__init__.py
@@ -23,3 +23,4 @@ from . import res_lang
 from . import sale_order_line
 from . import res_users
 from . import partner_alcyon_category
+from . import sale_order