+++ HISTORY.rst
@@ -35,6 +35,9 @@ Unreleased
 * [IMP] Disable tracking for sale order validation and improve the code to avoid to recompute quantities when it's not needed
 * [IMP] Fix a bug with the travis.yml file
 * ALCN-886 Add CNK code on products
+* ALCN-865: Web service create web sale order
+
+**Bugfixes**
 
 **Build**